+++ types/module/module_int_test.go
@@ -4,6 +4,7 @@ import (
 	"sort"
 	"testing"
 
+	"cosmossdk.io/core/appmodule"
 	"github.com/stretchr/testify/suite"
 )
 
@@ -17,7 +18,7 @@ type TestSuite struct {
 
 func (s *TestSuite) TestAssertNoForgottenModules() {
 	m := Manager{
-		Modules: map[string]interface{}{"a": nil, "b": nil},
+		Modules: map[string]appmodule.AppModule{"a": nil, "b": nil},
 	}
 	tcs := []struct {
 		name     string
@@ -42,7 +43,7 @@ func (s *TestSuite) TestAssertNoForgottenModules() {
 
 func (s *TestSuite) TestModuleNames() {
 	m := Manager{
-		Modules: map[string]interface{}{"a": nil, "b": nil},
+		Modules: map[string]appmodule.AppModule{"a": nil, "b": nil},
 	}
 	ms := m.ModuleNames()
 	sort.Strings(ms)